Refactor: Improve ingestion, executor, and project guidelines
This commit includes several refactorings and improvements across the DAM system.

- In `dam.core.executor`, the `_populate_results_if_needed` function now logs a warning instead of raising a `TypeError` when an `InformationRequest` is consumed by a method that cannot handle it. This makes the system more robust to unexpected states.

- The `ingest_archive_members_handler` in `dam_archive` has been updated to prevent the re-ingestion of individual parts of an already assembled split archive. It now logs a message and skips the part, ensuring that ingestion is only initiated from the master entity.

- The `dam.core.requests` module has been relocated to `dam.system_events.requests` to better reflect its purpose. All corresponding import statements have been updated.

- The project's `AGENTS.md` documentation has been updated to include a new guideline discouraging the re-export of symbols from `__init__.py` files to improve code clarity and maintainability.
